LangChain/requirements.txt
# Web Framework
fastapi>=0.110.1,<0.113
uvicorn>=0.25.0,<0.26.0
python-multipart>=0.0.9,<0.0.10
aiofiles>=23.1.0,<24.0.0

# Database
mysql-connector-python==8.2.0

# Message Queue
confluent-kafka==2.4.0

# Audio Processing
pyaudio==0.2.13
SpeechRecognition==3.10.0
gTTS==2.4.0
wave==0.0.2

# AI & Language Models
openai>=1.6.1, <2.0.0
langchain==0.3.6
langchain-openai==0.2.4
langchain-community==0.3.4
langchain-core<0.4.0, >=0.3.14
tiktoken==0.8.0


# Configuration & Settings
python-dotenv>=1.0.0,<2.0.0
pydantic>=2.7.4,<3.0.0
pydantic-settings>=2.1.0,<3.0.0

# Logging & Console
python-json-logger==2.0.7
colorama==0.4.6
keyboard==0.13.5

# HTTP & API
requests>=2.31.0,<3.0.0
httpx>=0.23.0,<1.0.0

# PDF Generation
fpdf==1.7.2

# Testing
pytest==8.1.1
pytest-asyncio==0.22.0